router_api/requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Django==3.0.6
djangorestframework==3.11.0
Faker==4.1.0
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
netmiko==3.1.1
numpy==1.18.4
pandas==1.0.3
paramiko==2.7.1
pkg-resources==0.0.0
pycparser==2.20
PyNaCl==1.3.0
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
roscraco==0.5.0
scp==0.13.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
textfsm==1.1.0
Werkzeug==1.0.1